+++ pkg/tracegen/options.go
@@ -1,4 +1,4 @@
-package client
+package tracegen
 
 import "time"
