+++ pkg/tracegen/subresource.go
@@ -1,4 +1,4 @@
-package client
+package tracegen
 
 import (
 	"context"
